FIX: StripeJS testing on slow connections
Previously these tests only worked if StripeJS responded in under two seconds (the Capybara default). A workaround to most of these cases is to wrap the `stripeResponseHandler` and keep track of when Stripe is active and then finished. For a request that also updates the Spree shop, we need to wait even longer.

The ultimate proper solution would be to stub out the Stripe API and build our own responses.